send encryption of fake battery data to phone for testing

+++ esp/esp_driver/network_adapter/main/tm/tm_ble_gatts_server.h
@@ -8,4 +8,4 @@ extern void tm_ble_gatts_server_init(void);
 extern void tm_ble_gatts_send_to_phone(uint8_t *data, int len);
 extern void tm_ble_gatts_kill_connection(void);
 extern uint32_t EndianConverter(uint32_t *byteArray);
-extern int serialize_data(uint8_t *buf, size_t pos, uint8_t *data, size_t len);
+extern size_t serialize_data(uint8_t *buf, size_t pos, uint8_t *data, size_t len);
